requirements.txt
openai>=1.30.0
anthropic>=0.25.0
langchain>=0.2.0
langchain-community>=0.2.0
faiss-cpu>=1.7.4
sentence-transformers>=2.7.0
python-dotenv>=1.0.0
